+++ DistroLauncher-Tests/DistroLauncher-Tests/DistroLauncher-Tests.vcxproj.filters
@@ -1,9 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="..\..\DistroLauncher\ApplicationStrategy.cpp">
-      <Filter>DistroSources</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\DistroLauncher\DistributionInfo.cpp">
       <Filter>DistroSources</Filter>
     </ClCompile>
